+++ SWLOR.Game.Server/Service/Contracts/ICustomEffectService.cs
@@ -16,7 +16,7 @@ namespace SWLOR.Game.Server.Service.Contracts
         void OnModuleLoad();
         void OnModuleEnter();
         void RemovePCCustomEffect(NWPlayer oPC, CustomEffectType customEffectType);
-        void RemovePCCustomEffect(NWPlayer oPC, long customEffectID);
+        void RemovePCCustomEffect(NWPlayer oPC, int customEffectID);
         int GetCustomEffectLevel(NWCreature creature, CustomEffectType customEffectType);
         void ApplyStance(NWPlayer player, CustomEffectType customEffect, PerkType perkType, int effectiveLevel, string data);
         bool RemoveStance(NWPlayer player, PCCustomEffect stanceEffect = null, bool sendMessage = true);
